App Routing - Consulting page
Affected Route: `/consulting`

Fixed: #3162

Screenshot:
TODO
